+++ src/main/java/gov/usgs/volcanoes/swarm/event/hypo71/Hypo71Settings.java
@@ -18,8 +18,9 @@ public class Hypo71Settings extends Properties {
   private static final long serialVersionUID = 2145784685211194449L;
   private static final String SETTINGS_FILENAME = "Hypo71.config";
   private static final String COMMENTS = "Hypo71 Settings";
+  public static final String KSING = "KSING";
+  public static final String ksingDefault = "0";
   public static final String TEST = "TEST";
-
   public static final double[] testDefault =
       {0.10f, 10.0f, 2.0f, 0.05f, 5.0f, 4.0f, -0.87f, 2.00f, 0.0035f, 100.0f, 8.0f, 0.5f, 1.0f};
   
@@ -47,6 +48,9 @@ public class Hypo71Settings extends Properties {
    * Load default properties in cases where settings file does not exist.
    */
   private void loadDefaultProperties() {
+    // KSING
+    this.setProperty(KSING, ksingDefault);
+    
     // Test Variables
     int i = 1;
     for (double t : testDefault) {
@@ -63,6 +67,16 @@ public class Hypo71Settings extends Properties {
   protected void save() throws IOException {
     FileWriter writer = new FileWriter(SETTINGS_FILENAME);
     store(writer, COMMENTS);
+  }
+  
+  /**
+   * Get KSING value for Control Card input.
+   * @return 0 if using original SINGLE subroutine;
+   *         1 if using modified SINGLE subroutine
+   */
+  protected int getKsing() {
+    int ksing = Integer.valueOf(this.getProperty(KSING));
+    return ksing;
   }
    
   /**